Refactor manager finances and dashboard components for improved pro-forma handling and UI updates
