--- src/app/models/pro-forma.model.ts
@@ -1,11 +1,0 @@
-export interface ProForma {
-    id: number;
-    orderId: number;
-    amount: number;
-    dueDate: Date;
-    status: 'pending' | 'paid' | 'overdue';
-    isPaid: boolean;
-    isCreated: boolean;
-    daysOverdue?: number;
-    notes?: string;
-}
